Fixup test deps
These were revealed broken by running in non-fast mode.
