+++ src/python/pants/backend/codegen/BUILD
@@ -10,5 +10,6 @@ python_library(
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/backend/codegen/targets:python',
     'src/python/pants/base:build_file_aliases',
+    'src/python/pants/goal:task_registrar',
   ]
 )
