+++ src/python/pants/option/BUILD
@@ -1,11 +1,12 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-
 python_library(
   name = 'option',
   sources = globs('*.py'),
   dependencies = [
     '3rdparty/python:argparse',
+    'src/python/pants/base:build_environment',
+    'src/python/pants/goal'
   ]
 )
