+++ tests/python/pants_test/backend/jvm/tasks/jvm_compile/BUILD
@@ -12,6 +12,7 @@ python_tests(
   name = 'jvm_fingerprint_strategy',
   sources = ['test_jvm_fingerprint_strategy.py'],
   dependencies = [
+    'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/tasks/jvm_compile:jvm_fingerprint_strategy',
     'tests/python/pants_test:base_test',
   ]
